tdf#130857 qt weld: Notify about spinbox combined value+text change
While QtInstanceEntry generally takes care of handling signals
for the spinbox's QLineEdit, this doesn't work when the value
is changed as a result of setting a new spinbox value (e.g.
by using the spinbox buttons), as the QLineEdit signals are blocked
then, see QAbstractSpinBoxPrivate::updateEdit in qtbase [1].
Therefore, connect the QDoubleSpinBox::textChanged signal
to the slot that calls signal_changed() instead to ensure
it gets called nonetheless, and disconnect from the other signal.

While at it, also add a SolarMutexGuard when calling the
signal.

This fixes the issue noticed with the "Go to Page"
dialog mentioned in previous commit

    Change-Id: I1f24cf3925e945ae78a9f1646535e08736cd8786

    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Nov 27 23:53:58 2024 +0100

        tdf#130857 qt weld: Support "Go to Page" dialog

as:

> One issue seen with the dialog is that by using the
> up arrow of the spinbox to increase the value, it
> is currently possible to increase the value
> beyond the maximum value (last page number),
> while this is not the case when typing a number
> into the box manually.
>
> This is because the GotoPageDlg::PageModifiedHdl
> handler currently only gets called for the
> latter case, not the former one, despite
>
>     Change-Id: Ie19bc852f4ceed0fa79565302975376db7126ea4
>     Author: Michael Weghorn <m.weghorn@posteo.de>
>     Date:   Wed Nov 27 22:53:55 2024 +0100
>
>         tdf#130857 qt weld: Also notify about programmatic text changes
>
> and will be addressed in a separate commit.

[1] https://code.qt.io/cgit/qt/qtbase.git/tree/src/widgets/widgets/qabstractspinbox.cpp?id=ced47a590aeb85953a16eaf362887f14c2815c45#n1790

Change-Id: Ifba9a0877442f9e84f0103d8aab202ae3583c5cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177451
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>