tdf#158979 Add static variable for last used merge option
A static variable has been added so that, when merging cells in Calc
where multiple cells have content in them, it will remember which of the three
merge options was last used in a given session and automatically select that
option the next time the merge cells dialog pops up. The default for the first
merge in a session remains the option to keep the content of hidden cells.

Change-Id: I35f6c01ffe9dff2756e3b9f5d530725e5d939db8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177455
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>