sw: layout: fix toggling Hidden Paragraphs that have follows
When a hidden paragraph is split across pages, there is nothing that
joins its follow (SwContentFrame::MakePrtArea() has code to do it but it
wasn't called even before the previous commit?) and when the follow is
formatted is doesn't move back.

So just join all follows asap in SwContentFrame::MakeAll().

(presumably regression from commit
 0c96119895b347f8eb5bb89f393351bd3c02b9f1 or one of its follow-ups)

Change-Id: Icecdf17c1c1e11a3b62c7e5bf2c6624f234c94f3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177500
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins