Related: tdf#163945 don't directly flush graphics with Skia/Metal
When dragging a selection box on an empty background in
Impress and only with Skia/Metal, the selection box
would not keep up with the pointer. The selection box
would repaint sporadically or not at all if the pointer
was dragged rapidly and the status bar was visible.

Apparently, flushing a graphics doesn't actually do much
of anything with Skia/Raster and Skia disabled so the
selection box repaints without any noticeable delay.
However, with Skia/Metal every flush of a graphics
creates and queues a new CAMetalLayer drawable. During
rapid dragging, this can lead to creating and queueing
up to 200 drawables per second leaving no spare time for
the Impress selection box painting timer to fire.

So with Skia/Metal, throttle the rate of flushing by
calling display on the view.

Also, with the reduced Skia/Metal flushing load from this
fix, the color conversion when drawing to an NSBox appears
to be no longer needed. The converted color appeared less
saturated, at least to me, so accept the (hopefully now
imperceptible) performance hit of drawing an NSBox with
native colors.

Change-Id: I55b4ab763bf20c6c2acad65587b703fc6f645264
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177530
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins
(cherry picked from commit c585b697b583fa0f8cdadeab594c31d270367ba7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177731
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>