tdf#88226 sd: fix cutting off the overflow text on the notes print page
This fix offers two options to preserve the overflowed text:
1: if the "Original size" option is selected for printing,
   the overflowed text will be displayed on a new page.
2: if the "Fit to Printable Area" option is selected for printing,
   notes will be scaled to fit within the available printable space.

The "Multiple sheets of paper" and "Tile sheet of paper" options are disabled for notes because these options are intended for slide printing and do not make much sense for printing notes.

The orientation for the notes print page has also been fixed.

Change-Id: I99e56cf9aed5c32764797469a8ea7f3b25053882
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177511
Reviewed-by: Nagy Tibor <tibor.nagy.extern@allotropia.de>
Tested-by: Jenkins