sw: layout: fix toggling Hidden Paragraphs not shrinking flys
There are 2 as-char flys with 2 hidden paragraphs each, and when turning
off show hidden paragraphs, the paragraphs shrink to height 0 but the
flys don't - this is because due to other hidden paragraphs earlier, the
position of the flys change and Shrink() is called before MakePos() and
then it uses the old position to check and that results in a negative
nRstHeight, so the upper doesn't actually shrink.

Fix this by always calling MakePos() before MakeValidZeroHeight().

(presumably regression from commit
 0c96119895b347f8eb5bb89f393351bd3c02b9f1 or one of its follow-ups)

Change-Id: I9a4ffadff02b6a2bc59c6a37aa15abec429e1068
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177524
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins