tdf#130857 qt weld: Implement basic tree view logic
Implement a small subset of the QtInstanceTreeView
methods - basically what is used by Calc's
"Sheets" -> "Show Sheets" dialog for which
support will be declared in an upcoming commit.

In QtBuilder, ignore "GtkCellRendererText" and
"GtkTreeSelection" objects, just as VclBuilder
does.
For now, also don't implement logic for handling
"GtkTreeViewColumn" objects yet, but add a warning
that this is still missing.

Use QStandardItemModel for the tree view's
model.

For those methods where not all possible
parameters are handled yet, add asserts
that will trigger when called with currently
still unsupported ones, so this can easily
be identified when working on support for
dialogs using more complex tree views in the
future.

QtInstanceTreeView::get_height_rows logic
to return the required height to show n
entries also still needs some thought, so the
method simply warns and returns 0 for now, which
means that the tree view might not be shown with
the ideal initial height, but resizing the dialog
is sufficient to work around that for now if needed.

Change-Id: I13d9bd4a742c29a3dfc2c17b6c1a13f9fe40270e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177541
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>