wsl-as-helper: look for perl modules in the perl that's actually used
i.e. strawberry-perl-portable for the installsets and openssl

The only module that needs to be installed (and only when building
installation sets) is Font::TTF::Font - but only on the windows side,
not in the wsl-container.

in the wsl-as-helper case there are three different versions of perl
involved:

  * one inside wsl, since autogen.sh is a perl script
  * one provided by git-bash - used for the majority of the build
    whenever a recipe uses perl
  * and strawberry-perl-portable for building openssl (since that is
    picky and needs one that handles the windows-paths a certain way)
    and for building the installation sets (because of similar
    assumptions in path mangling)

Change-Id: I8374749f21c7862f2e9e77d760077e836a6e9166
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177560
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>
(cherry picked from commit 6e93419f46924876da1f0c52d27c03968becf046)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177725