tdf#130857 qt weld: Notify about current tree view item change
When the current (focus) item in the tree view changes,
call weld::TreeView::signal_changed to notify about this.

While QTreeView doesn't have a signal for that itself,
its QItemSelectionModel does, so connect to that one.

With this in place, SvPasteObjectDialog::SelectHdl
now gets called when switching between items in the
"Past Special" dialog when using the qt6 VCL plugin
with native Qt widgets, as is the case with the gtk3
VCL plugin.

Change-Id: I5aab68055f711480d56b043252d15621f18c4919
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177573
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins