tdf#130857 qt weld: Support "Split Cells" dialog
Declare support for the "Split Cells" dialog that
can e.g. be triggered using the "Table" -> "Split Cells"
menu entry in Writer while the cursor is inside of a
table cell.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Change-Id: I5b994698ea46e9ede763169d21fe53527b3b369b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177597
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>