tdf#130857 qt weld: Support "Split Table" dialog
Declare support for the "Split Table" dialog that
can be triggered using the "Table" -> "Split Table"
menu entry in Writer while the cursor is inside of a
table cell.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Change-Id: Iae4855027de533ff7f9a4148f8101a56c281abcf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177598
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins