sc a11y: Don't suppress a11y events when setting new pos
Don't suppress a11y events when setting the new
position string in the combobox in Calc's formula
toolbar.

Suppressing events was added in

    commit 38fe580341a3f1e379edada08db7e1aebab258f5
    Author: Malte Timmermann [mt] <malte.timmermann@oracle.com>
    Date:   Thu Jan 20 16:34:26 2011 +0100

        accfixes: added more accessibility information and fixed tab orders in multiple dialogs (sc module)

for unclear reasons.

The only potential reason I can come up with could be
to avoid announcement by screen readers, but they
usually don't announce text/value changes to non-focused
controls anyway, and neither Orca on Linux nor NVDA
on Windows announcement changed with this change in place
when moving focus between cells in Calc.

Without this change in place, having the combobox's
"text" child selected in Accerciser's treeview
of LO's a11y hierarchy, the text shown in the "Text"
interface section would not automatically update to display
the newly selected location due to the missing
event when using the qt6 VCL plugin. With this change in
place, it is updated as expected.

Change-Id: I88fe0e6ec9bb693710e3f7ab0c6c2e46686be720
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177638
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins