lok: slideshow: handle text object in edit mode
This patch allows to render a text object in edit mode by decomposing a
TextHierarchyEditPrimitive2D instance in place of using TextEditDrawing.
That avoids to have artifacts displayed while playing the slideshow such as a
tiny rectangle around the edited text shape.
Moreover it allows to animate a single paragraph even when the related text
object is in edit mode.

Change-Id: I0c27dc0953dd1b0fb7ae70f41e569587f3a623ac
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177755
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>