Let StyleDispatcher take the state of StyleApply slot into account
Before, it only listened to ParaStyle, or CharStyle, etc. obtained
in MapFamilyToCommand; and that meant, that its actual command was
".uno:StyleApply?...", but StyleApply state was ignored. This meant
that all respective items under "Style" menu were enabled in read-
only mode, where StyleApply itself was inactive.

Listen to both. When StyleApply is disabled, that takes priority.

Change-Id: Ic087ba6fc6fe6d45b8c8ecc35494734cd8765bd4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/177971
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins